refactor: make service and util out of domain
