+++ src/main/java/lotto/domain/WinningRank.java
@@ -1,4 +1,4 @@
-package lotto.domain.repository;
+package lotto.domain;
 
 import java.util.Arrays;
 import java.util.List;
