+++ src/main/java/lotto/service/WinningRankStatistics.java
@@ -1,7 +1,7 @@
-package lotto.domain.service;
+package lotto.service;
 
-import lotto.domain.repository.WinningRank;
-import lotto.domain.repository.WinningRanks;
+import lotto.domain.WinningRank;
+import lotto.domain.WinningRanks;
 
 import java.math.BigDecimal;
 import java.math.RoundingMode;
@@ -9,7 +9,7 @@ import java.util.Arrays;
 import java.util.LinkedHashMap;
 import java.util.Map;
 
-import static lotto.domain.util.MoneyLottoMapper.LOTTO_PRICE;
+import static lotto.util.MoneyLottoMapper.LOTTO_PRICE;
 
 public class WinningRankStatistics {
     public static final int SCALE = 2;
