+++ src/test/java/lotto/domain/LotteryMachineTest.java
@@ -1,7 +1,6 @@
 package lotto.domain;
 
-import lotto.domain.repository.*;
-import lotto.domain.service.LotteryMachine;
+import lotto.service.LotteryMachine;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
