+++ src/test/java/lotto/domain/NumbersTest.java
@@ -1,6 +1,5 @@
 package lotto.domain;
 
-import lotto.domain.repository.Numbers;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
