+++ src/test/java/lotto/domain/WinningRankStatisticsTest.java
@@ -1,8 +1,6 @@
 package lotto.domain;
 
-import lotto.domain.repository.WinningRank;
-import lotto.domain.repository.WinningRanks;
-import lotto.domain.service.WinningRankStatistics;
+import lotto.service.WinningRankStatistics;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -12,7 +10,7 @@ import java.util.Arrays;
 import java.util.List;
 import java.util.Map;
 
-import static lotto.domain.util.MoneyLottoMapper.LOTTO_PRICE;
+import static lotto.util.MoneyLottoMapper.LOTTO_PRICE;
 import static org.assertj.core.api.Assertions.assertThat;
 
 class WinningRankStatisticsTest {
