+++ src/test/java/lotto/domain/WinningRankTest.java
@@ -1,6 +1,5 @@
 package lotto.domain;
 
-import lotto.domain.repository.WinningRank;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
